ranker/requirements.txt
celery==5.2.7
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
eventlet==0.33.1
filetype==1.1.0
fuzzyset2==0.2.0
gdown==4.4.0
gensim==4.2.0
huggingface-hub==0.8.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nltk==3.7
numpy==1.22.3
pandas==1.4.2
pdfminer==20191125
prettytable==3.3.0
pymongo==4.1.1
PyMuPDF==1.20.1
pyresparser==1.0.6
requests==2.28.1 
resume-parser==0.8.4
scikit-learn==1.1.1
sentence-transformers==2.2.1
six==1.16.0
spacy[transformers]
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
transformers==4.19.4
urllib3==1.26.9

